+++ packages/riverpod_analyzer_utils/lib/src/riverpod_ast/legacy_provider_declaration.dart
@@ -4,22 +4,28 @@ class LegacyProviderDependencies extends RiverpodAst {
   LegacyProviderDependencies._({
     required this.dependencies,
     required this.dependenciesNode,
+    required this.unit,
   });
 
-  static LegacyProviderDependencies? _parse(NamedExpression? dependenciesNode) {
+  static LegacyProviderDependencies? _parse(
+    NamedExpression? dependenciesNode, {
+    required CompilationUnit unit,
+  }) {
     if (dependenciesNode == null) return null;
 
     final value = dependenciesNode.expression;
 
     List<LegacyProviderDependency>? dependencies;
     if (value is ListLiteral) {
-      dependencies =
-          value.elements.map(LegacyProviderDependency._parse).toList();
+      dependencies = value.elements
+          .map((e) => LegacyProviderDependency._parse(e, unit: unit))
+          .toList();
     }
 
     final legacyProviderDependencies = LegacyProviderDependencies._(
       dependenciesNode: dependenciesNode,
       dependencies: dependencies,
+      unit: unit,
     );
     legacyProviderDependencies.dependencies?.forEach((element) {
       element._parent = legacyProviderDependencies;
@@ -28,6 +34,8 @@ class LegacyProviderDependencies extends RiverpodAst {
     return legacyProviderDependencies;
   }
 
+  @override
+  final CompilationUnit unit;
   final List<LegacyProviderDependency>? dependencies;
   final NamedExpression dependenciesNode;
 
@@ -52,20 +60,28 @@ class LegacyProviderDependency extends RiverpodAst
   LegacyProviderDependency._({
     required this.node,
     required this.provider,
+    required this.unit,
   });
 
-  factory LegacyProviderDependency._parse(CollectionElement node) {
-    final provider =
-        node.cast<Expression>().let(ProviderListenableExpression._parse);
+  factory LegacyProviderDependency._parse(
+    CollectionElement node, {
+    required CompilationUnit unit,
+  }) {
+    final provider = node
+        .cast<Expression>()
+        .let((e) => ProviderListenableExpression._parse(e, unit: unit));
 
     final legacyProviderDependency = LegacyProviderDependency._(
       node: node,
       provider: provider,
+      unit: unit,
     );
     provider?._parent = legacyProviderDependency;
     return legacyProviderDependency;
   }
 
+  @override
+  final CompilationUnit unit;
   final CollectionElement node;
   final ProviderListenableExpression? provider;
 
@@ -93,12 +109,14 @@ class LegacyProviderDeclaration extends RiverpodAst
     required this.autoDisposeModifier,
     required this.familyModifier,
     required this.dependencies,
+    required this.unit,
   });
 
   static LegacyProviderDeclaration? _parse(
     VariableDeclaration node,
-    _ParseRefInvocationMixin parent,
-  ) {
+    _ParseRefInvocationMixin parent, {
+    required CompilationUnit unit,
+  }) {
     final element = node.declaredElement;
     if (element == null) return null;
 
@@ -163,7 +181,10 @@ class LegacyProviderDeclaration extends RiverpodAst
     final dependenciesElement = arguments
         .namedArguments()
         .firstWhereOrNull((e) => e.name.label.name == 'dependencies');
-    final dependencies = LegacyProviderDependencies._parse(dependenciesElement);
+    final dependencies = LegacyProviderDependencies._parse(
+      dependenciesElement,
+      unit: unit,
+    );
 
     final legacyProviderDeclaration = LegacyProviderDeclaration._(
       name: node.name,
@@ -176,6 +197,7 @@ class LegacyProviderDeclaration extends RiverpodAst
       autoDisposeModifier: autoDisposeModifier,
       familyModifier: familyModifier,
       dependencies: dependencies,
+      unit: unit,
     );
 
     dependencies?._parent = legacyProviderDeclaration;
@@ -183,11 +205,15 @@ class LegacyProviderDeclaration extends RiverpodAst
       _LegacyRefInvocationVisitor(
         legacyProviderDeclaration,
         parent,
+        unit: unit,
       ),
     );
 
     return legacyProviderDeclaration;
   }
+
+  @override
+  final CompilationUnit unit;
 
   final LegacyProviderDependencies? dependencies;
 
@@ -225,7 +251,14 @@ class LegacyProviderDeclaration extends RiverpodAst
 
 class _LegacyRefInvocationVisitor extends RecursiveAstVisitor<void>
     with _ParseRefInvocationMixin {
-  _LegacyRefInvocationVisitor(this.declaration, this.parent);
+  _LegacyRefInvocationVisitor(
+    this.declaration,
+    this.parent, {
+    required this.unit,
+  });
+
+  @override
+  final CompilationUnit unit;
 
   final LegacyProviderDeclaration declaration;
   final _ParseRefInvocationMixin parent;