+++ packages/riverpod_analyzer_utils/lib/src/riverpod_ast/provider_declaration.dart
@@ -1,6 +1,8 @@
 part of '../riverpod_ast.dart';
 
 abstract class ProviderDeclaration extends RiverpodAst {
+  @override
+  CompilationUnit get unit;
   Token get name;
   AnnotatedNode get node;
   ProviderDeclarationElement get providerElement;